Move Tag routes to `route`
