+++ src/main/scala/com/hhandoko/realworld/Server.scala
@@ -14,8 +14,8 @@ import com.hhandoko.realworld.article.ArticleService
 import com.hhandoko.realworld.auth.{AuthRoutes, AuthService, RequestAuthenticator}
 import com.hhandoko.realworld.config.{Config, DbConfig}
 import com.hhandoko.realworld.profile.ProfileService
-import com.hhandoko.realworld.route.{ArticleRoutes, ProfileRoutes}
-import com.hhandoko.realworld.tag.{TagRoutes, TagService}
+import com.hhandoko.realworld.route.{ArticleRoutes, ProfileRoutes, TagRoutes}
+import com.hhandoko.realworld.tag.TagService
 import com.hhandoko.realworld.user.{UserRoutes, UserService}
 
 object Server {
