+++ src/main/scala/com/hhandoko/realworld/route/TagRoutes.scala
@@ -1,19 +1,21 @@
-package com.hhandoko.realworld.tag
+package com.hhandoko.realworld.route
 
 import cats.Applicative
 import cats.effect.{ContextShift, Sync}
 import cats.implicits._
 import io.circe.{Encoder, Json}
-import org.http4s.circe.jsonEncoderOf
+import org.http4s.circe._
 import org.http4s.dsl.Http4sDsl
 import org.http4s.{EntityEncoder, HttpRoutes}
 
 import com.hhandoko.realworld.core.Tag
+import com.hhandoko.realworld.tag.TagService
 
 object TagRoutes {
 
-  def apply[F[_]: Sync: ContextShift](tagService: TagService[F]): HttpRoutes[F] = {
-    object dsl extends Http4sDsl[F]; import dsl._
+  def apply[F[_] : Sync : ContextShift](tagService: TagService[F]): HttpRoutes[F] = {
+    object dsl extends Http4sDsl[F];
+    import dsl._
 
     HttpRoutes.of[F] {
       case GET -> Root / "api" / "tags" =>
@@ -31,8 +33,7 @@ object TagRoutes {
       "tags" -> Json.fromValues(r.tags.map(_.value).map(Json.fromString))
     )
 
-    implicit def entityEncoder[F[_]: Applicative]: EntityEncoder[F, AllTagsResponse] =
+    implicit def entityEncoder[F[_] : Applicative]: EntityEncoder[F, AllTagsResponse] =
       jsonEncoderOf[F, AllTagsResponse]
   }
-
 }
